requirements.txt
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Django==1.11
django-extensions==2.1.3
gunicorn==19.9.0
idna==2.7
pycparser==2.19
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
jsonfield==2.0.2
psycopg2-binary==2.7.7
djangorestframework==3.8.2
